Add max_sequence_idle_microseconds
This setting ensures that inference server resources are freed after a
given timeout even if a sequence doesn't send an end marker. Without
this timeout malicious or faulty sequences could DOS the server.

+++ src/core/dynamic_batch_scheduler.cc
@@ -72,8 +72,7 @@ DynamicBatchScheduler::DynamicBatchScheduler(
     }
 
     pending_batch_delay_ns_ =
-        (uint64_t)config.dynamic_batching().max_queue_delay_microseconds() *
-        1000;
+        config.dynamic_batching().max_queue_delay_microseconds() * 1000;
   }
 }
